refactor(quality): extract Codex Quality client into codex-quality.ts
Task 16 subagent didn't know src/utils/quality.ts is occupied by
legacy AutoTuner/KPI client (253 lines, consumed by QualityPanel.tsx
and existing useQuality hook). It appended new Codex code there as a
named section. Two unrelated API surfaces in one file is confusing
and dangerous (import collisions, accidental edits).

Move fetchQualityData / fetchAnnotations / forceQualityRefresh /
createAnnotation / deleteAnnotation + URL helpers + declare global
into a NEW file src/utils/codex-quality.ts. Revert quality.ts to
its original 253-line state.

All callers from now on import from "./codex-quality" — this avoids
future merge conflicts with the legacy AutoTuner client. Same plan
will apply to Task 17 (useQuality.ts → useCodexQuality.ts).
